NaxcaFlask/NaxcaFlask/requirements.txt
flask-security
flask-sqlalchemy
psycopg2